fix(build): Vite doesn't support tsc path aliases
TIL: Vite out of the box does not support resolving modules using path
aliases defined in a user's `tsconfig.json` file. This is an Astro
feature. I thought since Astro is built on top of Vite, this was a
feature that was carried over. This is incorrect. That took a lil while
for me to figure out...