feat: show active leads with scores
Keeping it simple by showing scores as a progress bar, intentionally not
showing the number value.

We haven't made a page to create new leads yet, but I am planing on
having a range input for quickly entering a value for the score. I want
to have as few inputs as possible for creating a lead and have them be
extremely quick to operate.